moved track rest service to separate package
