+++ application/src/main/java/com/musicforall/services/DbPopulateService.java
@@ -8,6 +8,7 @@ import com.musicforall.model.Track;
 import com.musicforall.model.User;
 import com.musicforall.services.follower.FollowerService;
 import com.musicforall.services.playlist.PlaylistService;
+import com.musicforall.services.track.TrackService;
 import com.musicforall.services.user.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
